fix:ドラッグから選択に変えたのでメソッドや変数名などを修正
